+++ packages/core/cypress/support/index.js
@@ -1,3 +1,3 @@
+import 'cypress-axe'
 import './commands'
 import '@testing-library/cypress/add-commands'
-import 'cypress-axe'